fix(Tests): fix on Tests + Subscriber updated
