--- src/Subscriber/Interfaces/RegisterCredentialsSubscriberInterface.php
@@ -1,29 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-/*
- * This file is part of the MarketReminder project.
- *
- * (c) Guillaume Loulier <contact@guillaumeloulier.fr>
- *
- * For the full copyright and license information, please view the LICENSE
- * file that was distributed with this source code.
- */
-
-namespace App\Subscriber\Interfaces;
-
-use Symfony\Component\Form\FormEvent;
-
-/**
- * Interface RegisterCredentialsSubscriberInterface.
- *
- * @author Guillaume Loulier <contact@guillaumeloulier.fr>
- */
-interface RegisterCredentialsSubscriberInterface
-{
-    /**
-     * @param FormEvent $event
-     */
-    public function checkCredentials(FormEvent $event): void;
-}
